==> src/Col2/45_soc_download_matrix.ipynb <==
# Download soil organic carbon design matrix asset (data frame) from GEE

Alessandro Samuel-Rosa, Taciara Zborowski Horst

2025-06-19

This script exports the soil Soil Organic Carbon (SOC) stock design matrix from Google Earth Engine
(GEE) to Google Drive as a CSV file in the "Earth Engine Exports" directory.

License: MIT

# Load required libraries
import ee
import ee.batch

# Initialize the Earth Engine API
# ee.Authenticate()
ee.Initialize(project='mapbiomas-solos-workspace')

# Define the path to the FeatureCollection directory
fc_dir = 'projects/mapbiomas-workspace/SOLOS/AMOSTRAS/MATRIZES/soil_organic_carbon/'

# Define the output Google Drive folder
gdrive_folder = 'Earth Engine Exports'

## Model 1: Amazon

# matriz-collection2_MODEL1_v2
fc_name = 'matriz-collection2_MODEL1_v2'
fc_path = fc_dir + fc_name
# Export the FeatureCollection to Google Drive
fc = ee.FeatureCollection(fc_path)
task = ee.batch.Export.table.toDrive(
    collection=fc,
    description=f'ASR Export {fc_name} to Drive',
    folder=gdrive_folder,
    fileNamePrefix=fc_name,
    fileFormat='CSV'
)
task.start()
print(f"Export task started. Check Google Drive for file.")
print(task.status())

## Model 2: Cerrado, Pantanal, Caatinga, and Atlantic Forest

# matriz-collection2_MODEL2_v2
fc_name = 'matriz-collection2_MODEL2_v2'
fc_path = fc_dir + fc_name
# Export the FeatureCollection to Google Drive
fc = ee.FeatureCollection(fc_path)
task = ee.batch.Export.table.toDrive(
    collection=fc,
    description=f'ASR Export {fc_name} to Drive',
    folder=gdrive_folder,
    fileNamePrefix=fc_name,
    fileFormat='CSV'
)
task.start()
print(f"Export task started. Check Google Drive for file.")
print(task.status())

## Model 3: Pampa

# matriz-collection2_MODEL3_v2
fc_name = 'matriz-collection2_MODEL3_v2'
fc_path = fc_dir + fc_name
# Export the FeatureCollection to Google Drive
fc = ee.FeatureCollection(fc_path)
task = ee.batch.Export.table.toDrive(
    collection=fc,
    description=f'ASR Export {fc_name} to Drive',
    folder=gdrive_folder,
    fileNamePrefix=fc_name,
    fileFormat='CSV'
)
task.start()
print(f"Export task started. Check Google Drive for file.")
print(task.status())
